+++ src/android/notification/Builder.java
@@ -364,17 +364,8 @@ public final class Builder {
 
         int reqCode = random.nextInt();
 
-        PendingIntent deleteIntent;
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.S) {
-            deleteIntent = PendingIntent.getBroadcast(
-                context, reqCode, intent, PendingIntent.FLAG_IMMUTABLE | FLAG_UPDATE_CURRENT);
-        } else {
-            deleteIntent  = PendingIntent.getBroadcast(
-                context, reqCode, intent, FLAG_UPDATE_CURRENT);
-
-        }
-
-        builder.setDeleteIntent(deleteIntent);
+        builder.setDeleteIntent(PendingIntent.getBroadcast(
+            context, reqCode, intent, PendingIntent.FLAG_IMMUTABLE | FLAG_UPDATE_CURRENT));
     }
 
     /**
@@ -399,17 +390,9 @@ public final class Builder {
         }
 
         int reqCode = random.nextInt();
-
-        PendingIntent contentIntent ;
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.S) {
-            contentIntent = PendingIntent.getActivity(
-                context, reqCode, intent, PendingIntent.FLAG_IMMUTABLE | FLAG_UPDATE_CURRENT);
-        } else {
-            contentIntent  = PendingIntent.getActivity(
-                context, reqCode, intent, FLAG_UPDATE_CURRENT);
-
-        }
-        builder.setContentIntent(contentIntent);
+        
+        builder.setContentIntent(PendingIntent.getActivity(
+            context, reqCode, intent, PendingIntent.FLAG_IMMUTABLE | FLAG_UPDATE_CURRENT));
     }
 
     /**
@@ -456,13 +439,7 @@ public final class Builder {
 
         int reqCode = random.nextInt();
 
-        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.S) {
-            return PendingIntent.getActivity(
-                context, reqCode, intent, PendingIntent.FLAG_IMMUTABLE | FLAG_UPDATE_CURRENT);
-        } else {
-            return PendingIntent.getActivity(
-                context, reqCode, intent, FLAG_UPDATE_CURRENT);
-        }
+        return PendingIntent.getActivity(context, reqCode, intent, PendingIntent.FLAG_IMMUTABLE | FLAG_UPDATE_CURRENT);
     }
 
     /**